# pva.SuperV.Api/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SuperV API",
    "description": "API for accessing SuperV projects.",
    "version": "v1"
  },
  "paths": {
...
    "/instances/{projectId}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets the list of instances from a project",
        "description": "Gets the list of instances from a project",
        "operationId": "GetInstances",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldDefinitionModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Creates an instance with a class of a project",
        "description": "Creates an instance with a class of a project",
        "operationId": "CreateInstance",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Instance creation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{projectId}/{instanceName}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets an instance of a project by its name",
        "description": "Gets an instance of a project by its name",
        "operationId": "GetInstance",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceName",
            "in": "path",
            "description": "Name of instance",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Deletes an instance from a project by its name",
        "description": "Deletes an instance from a project by its name",
        "operationId": "DeleteInstance",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceName",
            "in": "path",
            "description": "Name of instance",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{projectId}/{instanceName}/{fieldName}/value": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets the field of an instance from a project by its name",
        "description": "Gets the field of an instance from a project by its name",
        "operationId": "GetField",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceName",
            "in": "path",
            "description": "Name of instance",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldName",
            "in": "path",
            "description": "Name of field",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Updates the value of a field of an instance from a project by its name",
        "description": "Updates the value of a field of an instance from a project by its name",
        "operationId": "UpdateFieldValue",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceName",
            "in": "path",
            "description": "Name of instance",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldName",
            "in": "path",
            "description": "Name of field",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Value of field",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldValueModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FieldDefinitionModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/FieldDefinitionModelBoolFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelDateTimeFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelDoubleFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelFloatFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelIntFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelLongFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelShortFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelStringFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelTimeSpanFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUintFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUlongFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUshortFieldDefinitionModel"
          }
        ],
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "BoolFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelBoolFieldDefinitionModel",
            "DateTimeFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelDateTimeFieldDefinitionModel",
            "DoubleFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelDoubleFieldDefinitionModel",
            "FloatFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelFloatFieldDefinitionModel",
            "IntFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelIntFieldDefinitionModel",
            "LongFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelLongFieldDefinitionModel",
            "ShortFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelShortFieldDefinitionModel",
            "StringFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelStringFieldDefinitionModel",
            "TimeSpanFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelTimeSpanFieldDefinitionModel",
            "UintFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUintFieldDefinitionModel",
            "UlongFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUlongFieldDefinitionModel",
            "UshortFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUshortFieldDefinitionModel"
          }
        }
      },
      "FieldDefinitionModelBoolFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "BoolFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "boolean",
            "description": "Default value"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelDateTimeFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DateTimeFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "string",
            "description": "Default value",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelDoubleFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DoubleFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "number",
            "description": "Default value",
            "format": "double"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelFloatFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "FloatFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "number",
            "description": "Default value",
            "format": "float"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelIntFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "IntFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelLongFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "LongFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelShortFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "ShortFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int16"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelStringFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "StringFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "string",
            "description": "Default value",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelTimeSpanFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "TimeSpanFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string",
            "description": "Default value"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUintFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UintFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint32"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUlongFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UlongFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint64"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUshortFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UshortFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint16"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
...
      "FieldModel": {
        "required": [
          "name",
          "type",
          "fieldValue"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "fieldValue": {
            "$ref": "#/components/schemas/FieldValueModel"
          }
        }
      },
      "FieldValueModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/FieldValueModelBoolFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelDateTimeFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelDoubleFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelFloatFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelIntFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelLongFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelShortFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelStringFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelTimeSpanFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelUintFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelUlongFieldValueModel"
          },
          {
            "$ref": "#/components/schemas/FieldValueModelUshortFieldValueModel"
          }
        ],
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "BoolFieldValueModel": "#/components/schemas/FieldValueModelBoolFieldValueModel",
            "DateTimeFieldValueModel": "#/components/schemas/FieldValueModelDateTimeFieldValueModel",
            "DoubleFieldValueModel": "#/components/schemas/FieldValueModelDoubleFieldValueModel",
            "FloatFieldValueModel": "#/components/schemas/FieldValueModelFloatFieldValueModel",
            "IntFieldValueModel": "#/components/schemas/FieldValueModelIntFieldValueModel",
            "LongFieldValueModel": "#/components/schemas/FieldValueModelLongFieldValueModel",
            "ShortFieldValueModel": "#/components/schemas/FieldValueModelShortFieldValueModel",
            "StringFieldValueModel": "#/components/schemas/FieldValueModelStringFieldValueModel",
            "TimeSpanFieldValueModel": "#/components/schemas/FieldValueModelTimeSpanFieldValueModel",
            "UintFieldValueModel": "#/components/schemas/FieldValueModelUintFieldValueModel",
            "UlongFieldValueModel": "#/components/schemas/FieldValueModelUlongFieldValueModel",
            "UshortFieldValueModel": "#/components/schemas/FieldValueModelUshortFieldValueModel"
          }
        }
      },
      "FieldValueModelBoolFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "BoolFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "boolean",
            "description": "Field value."
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelDateTimeFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DateTimeFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "string",
            "description": "Field value.",
            "format": "date-time"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelDoubleFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DoubleFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "number",
            "description": "Field value.",
            "format": "double"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelFloatFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "FloatFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "number",
            "description": "Field value.",
            "format": "float"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelIntFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "IntFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "integer",
            "description": "Field value.",
            "format": "int32"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelLongFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "LongFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "integer",
            "description": "Field value.",
            "format": "int64"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "FieldValueModelShortFieldValueModel": {
        "required": [
          "value",
          "quality",
          "timestamp"
        ],
        "properties": {
          "$type": {
            "enum": [
              "ShortFieldValueModel"
            ],
            "type": "string"
          },
          "value": {
            "type": "integer",
            "description": "Field value.",
            "format": "int16"
          },
          "quality": {
            "type": "integer",
            "description": "Field value quality level.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Field value timestamp.",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "InstanceModel": {
        "required": [
          "name",
          "className",
          "fields"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Instance name."
          },
          "className": {
            "type": "string",
            "description": "Class name of instance."
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldModel"
            },
            "description": "Instance field values."
          }
        }
      },
...
    }
  },
...
}